requirements.txt
# =============================================================
# requirements.txt
# [RUBRIC MP4: Documentation & Reproducibility — 3 marks]
# Install with: pip install -r requirements.txt
# Python version: 3.9+
# =============================================================

# Deep Learning
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0

# Data & Numerics
numpy==1.24.4
pandas==2.1.4
scipy==1.11.4
scikit-learn==1.3.2

# Image Quality Metrics
scikit-image==0.22.0

# Computer Vision
opencv-python==4.8.1.78
Pillow==10.1.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Progress bars
tqdm==4.66.1

# Streamlit Deployment
streamlit==1.29.0

# Experiment tracking (optional but good practice)
# mlflow==2.9.2
